# src/role-has-permission/role-has-permission.controller.ts
import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
import { RoleHasPermissionService } from './role-has-permission.service';
import { CreateRoleHasPermissionDto } from './dto/create-role-has-permission.dto';
import { UpdateRoleHasPermissionDto } from './dto/update-role-has-permission.dto';

@Controller('role-has-permission')
export class RoleHasPermissionController {
  constructor(private readonly roleHasPermissionService: RoleHasPermissionService) {}

  @Post()
  create(@Body() createRoleHasPermissionDto: CreateRoleHasPermissionDto) {
    return this.roleHasPermissionService.create(createRoleHasPermissionDto);
  }

  @Get()
  findAll() {
    return this.roleHasPermissionService.findAll();
  }

  @Get(':id')
  findOne(@Param('id') id: string) {
    return this.roleHasPermissionService.findOne(+id);
  }

  @Patch(':id')
  update(@Param('id') id: string, @Body() updateRoleHasPermissionDto: UpdateRoleHasPermissionDto) {
    return this.roleHasPermissionService.update(+id, updateRoleHasPermissionDto);
  }

  @Delete(':id')
  remove(@Param('id') id: string) {
    return this.roleHasPermissionService.remove(+id);
  }
}